requirements.txt
# required Python >=3.11,<3.12
Flask==3.0.0
requests==2.31.0
Flask-SQLAlchemy==3.0.2
pandas==2.2.2
python-dotenv==1.0.0
pymysql==1.1.1
